pyproject.toml
[project]
name = "simple-rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.21",
    "fastapi>=0.116.1",
    "groq>=0.31.1",
    "pydantic>=2.11.8",
    "pypdf>=6.0.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.49.1",
    "uvicorn>=0.35.0",
]
